requirements.txt
python-telegram-bot
pandas
requests
matplotlib
python-dotenv
